\name{ABIdata}
\docType{data}
\alias{ABIdata}
\title{Gene Expression Data Matrix}
\description{93 ABI Human Genome Survey Microarray v2 samples, from a survey of gene expression in 31 different tissues}
\usage{data(GESTr)}
\format{Numeric matrix containing log2 normalised expression data for first 1000 genes across 93 samples.}
\source{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE7905}
\keyword{datasets}
