\name{fitGMM}
\alias{fitGMM}
\title{Fitting Gaussian Mixture Models}
\description{Fit Gaussian Mixture Model to a vector of gene expression data values.}
\usage{fitGMM(exprVals,RDparameters,rejectNull=0.05)}
\arguments{
	\item{exprVals}{A numeric vector of expression values for one gene}
	\item{RDparameters}{A list of Rocke-Durbin error model parameters, generated by the \code{fitRockeDurbin} function}
	\item{rejectNull}{Numeric value specifying significance level for rejecting null hypothesis that the expression values arise from a single value and expected errors}
}
\details{Uses \code{Mclust} to fit a Gaussian Mixture Model (with an unspecified number of components) to a vector of expression values for one gene across a dataset. Required to create models used by \code{GESTr}.}
\value{An object of class \code{Mclust}}
\author{Ed Curry \email{e.curry@imperial.ac.uk}}
\examples{
### internal function ###
}
